feat: DAO 테이블 책임 분리

+++ src/main/java/roomescape/domain/reservation/service/ReservationTimeService.java
@@ -2,18 +2,23 @@ package roomescape.domain.reservation.service;
 
 import java.util.List;
 import org.springframework.stereotype.Service;
+import roomescape.common.exception.AlreadyInUseException;
 import roomescape.domain.reservation.dto.ReservationTimeRequest;
 import roomescape.domain.reservation.dto.ReservationTimeResponse;
 import roomescape.domain.reservation.entity.ReservationTime;
+import roomescape.domain.reservation.repository.ReservationRepository;
 import roomescape.domain.reservation.repository.ReservationTimeRepository;
 
 @Service
 public class ReservationTimeService {
 
     private final ReservationTimeRepository reservationTimeRepository;
+    private final ReservationRepository reservationRepository;
 
-    public ReservationTimeService(ReservationTimeRepository reservationTimeRepository) {
+    public ReservationTimeService(ReservationTimeRepository reservationTimeRepository,
+                                  ReservationRepository reservationRepository) {
         this.reservationTimeRepository = reservationTimeRepository;
+        this.reservationRepository = reservationRepository;
     }
 
     public List<ReservationTimeResponse> getAll() {
@@ -32,8 +37,10 @@ public class ReservationTimeService {
         return ReservationTimeResponse.from(saved);
     }
 
-    public void delete(Long id){
-        // TODO 이미 예약에서 사용할 경우 AlreadyInuse 예외 발생
+    public void delete(Long id) {
+        if (reservationRepository.existsByTimeId(id)) {
+            throw new AlreadyInUseException("Reservation is already in use");
+        }
 
         reservationTimeRepository.deleteById(id);
     }
